multiple tests in examples folder
